backend/requirements.txt

fastapi==0.100.0
uvicorn[standard]==0.23.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[dcrypt]==1.7.4
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
pyjwt==2.8.0
email-validator==2.0.0.post2
bcrypt==4.0.1